test: Task 문서화 테스트 final 키워드 추가 - 문서화 테스트 명 변경
