+++ dailyge-api/src/test/java/project/dailyge/app/test/task/documentationtest/TaskRegisterDocumentationTest.java
@@ -8,48 +8,39 @@ import org.springframework.beans.factory.annotation.Autowired;
 import static org.springframework.http.MediaType.APPLICATION_JSON_VALUE;
 import static org.springframework.restdocs.restassured.RestAssuredRestDocumentation.document;
 import project.dailyge.app.common.DatabaseTestBase;
-import project.dailyge.app.common.auth.DailygeUser;
 import project.dailyge.app.core.task.dto.requesst.TaskRegisterRequest;
-import project.dailyge.app.core.task.facade.TaskFacade;
 import project.dailyge.app.core.user.application.UserWriteUseCase;
-import static project.dailyge.app.fixture.user.UserFixture.createUserJpaEntity;
+import project.dailyge.app.fixture.user.UserFixture;
 import static project.dailyge.app.test.task.documentationtest.snippet.TaskSnippet.TASK_AUTHORIZATION_HEADER;
-import static project.dailyge.app.test.task.documentationtest.snippet.TaskSnippet.TASK_DETAIL_SEARCH_PATH_PARAMETER_SNIPPET;
-import static project.dailyge.app.test.task.documentationtest.snippet.TaskSnippet.TASK_DETAIL_SEARCH_RESPONSE_SNIPPET;
-import project.dailyge.domain.task.TaskJpaEntity;
+import static project.dailyge.app.test.task.documentationtest.snippet.TaskSnippet.TASK_CREATE_REQUEST_SNIPPET;
+import static project.dailyge.app.test.task.documentationtest.snippet.TaskSnippet.TASK_CREATE_RESPONSE_SNIPPET;
 import project.dailyge.domain.user.UserJpaEntity;
 
-@DisplayName("[DocumentationTest] 할 일 저장 문서화 테스트")
-class TaskDetailSearchDatabaseTest extends DatabaseTestBase {
+@DisplayName("[DocumentationTest] 할 일 등록 문서화 테스트")
+class TaskRegisterDocumentationTest extends DatabaseTestBase {
 
     @Autowired
     private UserWriteUseCase userWriteUseCase;
 
-    @Autowired
-    private TaskFacade taskFacade;
-
     @Test
-    @DisplayName("Task가 존재하면 200 OK 응답을 받는다.")
-    void whenTaskExistsThenStatusCodeShouldBe200_OK() {
-        final UserJpaEntity newUser = userWriteUseCase.save(createUserJpaEntity());
+    @DisplayName("할 일을 등록하면 201 Created 응답을 받는다.")
+    void whenSaveTaskThenStatusCodeShouldBe_201() throws Exception {
+        final UserJpaEntity newUser = userWriteUseCase.save(UserFixture.createUserJpaEntity());
         final TaskRegisterRequest request = new TaskRegisterRequest("주간 미팅", "Backend 팀과 Api 스펙 정의", now());
-        final DailygeUser dailygeUser = new DailygeUser(newUser);
-        final TaskJpaEntity newTaskId = taskFacade.save(request.toEntity(dailygeUser));
 
         given(this.specification)
             .filter(document(IDENTIFIER,
                 TASK_AUTHORIZATION_HEADER,
-                TASK_DETAIL_SEARCH_PATH_PARAMETER_SNIPPET,
-                TASK_DETAIL_SEARCH_RESPONSE_SNIPPET
+                TASK_CREATE_REQUEST_SNIPPET,
+                TASK_CREATE_RESPONSE_SNIPPET
             ))
             .contentType(APPLICATION_JSON_VALUE)
             .header(AUTHORIZATION, "Token")
             .header(USER_ID_KEY, newUser.getId())
+            .body(objectMapper.writeValueAsString(request))
             .when()
-            .get("/api/tasks/{taskId}", newTaskId.getId())
+            .post("/api/tasks")
             .then()
-            .statusCode(200)
-            .log()
-            .all();
+            .statusCode(201);
     }
 }
